+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/jobscheduler/jobs/CatalogSyncJob.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024, Wazuh Inc.
+ * Copyright (C) 2024-2026, Wazuh Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -28,6 +28,7 @@ import java.util.concurrent.Semaphore;
 
 import com.wazuh.contentmanager.cti.catalog.index.ConsumersIndex;
 import com.wazuh.contentmanager.cti.catalog.synchronizer.AbstractConsumerSynchronizer;
+import com.wazuh.contentmanager.cti.catalog.synchronizer.IocConsumerSynchronizer;
 import com.wazuh.contentmanager.cti.catalog.synchronizer.UnifiedConsumerSynchronizer;
 import com.wazuh.contentmanager.jobscheduler.JobExecutor;
 
@@ -64,7 +65,9 @@ public class CatalogSyncJob implements JobExecutor {
             ThreadPool threadPool) {
         this.threadPool = threadPool;
         this.synchronizers =
-            List.of(new UnifiedConsumerSynchronizer(client, consumersIndex, environment));
+                List.of(
+                        new UnifiedConsumerSynchronizer(client, consumersIndex, environment),
+                        new IocConsumerSynchronizer(client, consumersIndex, environment));
     }
 
     /**