+++ CHANGELOG.md
@@ -57,6 +57,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add kvdb management endpoints [(#752)](https://github.com/wazuh/wazuh-indexer-plugins/pull/752)
 - Add draft policy update endpoint [(#758)](https://github.com/wazuh/wazuh-indexer-plugins/pull/758) [(#762)](https://github.com/wazuh/wazuh-indexer-plugins/pull/762)
 - Create integrations with no rules in the SAP [(#765)](https://github.com/wazuh/wazuh-indexer-plugins/pull/765)
+- Download IoC from CTI API [(#779)](https://github.com/wazuh/wazuh-indexer-plugins/pull/779)
 - Create default policies for user content spaces [(#767)](https://github.com/wazuh/wazuh-indexer-plugins/pull/767)
 - Add engine filters index [(#798)](https://github.com/wazuh/wazuh-indexer-plugins/pull/798)
 
@@ -122,6 +123,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fix integrations not being created under the correct category [#705](https://github.com/wazuh/wazuh-indexer-plugins/pull/705)
 - Set `GSON` object-to-number policy to `LONG_OR_DOUBLE` in `Resources` model [#721](https://github.com/wazuh/wazuh-indexer-plugins/pull/721)
 - Fix deserialization bug [(#737)](https://github.com/wazuh/wazuh-indexer-plugins/pull/737)
+- Fix offset value used when updating consumer state [(#802)](https://github.com/wazuh/wazuh-indexer-plugins/pull/802)
 
 ### Security
 - Reduce risk of GITHUB_TOKEN exposure [(#484)](https://github.com/wazuh/wazuh-indexer-plugins/pull/484)